1. 增加地图村庄 2. 增加人物和动物npc 3. 怪物无法进入村庄 4. npc,人物,怪物增加底部阴影

+++ app/src/main/java/com/game/dream/ui/Projectile.java
@@ -120,7 +120,7 @@ public class Projectile {
     /**
      * Draw the projectile
      */
-    public void draw(Canvas canvas, int offsetX, int offsetY) {
+    public void draw(Canvas canvas, float offsetX, float offsetY) {
         if (!isActive) return;
 
         Paint paint = new Paint();
